Module Specifications

module Assign00 : sig ... end

module Stdlib320 : sig ... end

The following is the standard library that we will use for CAS CS 320: Concepts of Programming Languages at Boston University in the Fall 2024 semester (Note. This is the first time we're doing this, so it's something of an experiment).