Module Assign01.Assign00_01

val pow : int -> int -> int

Perhaps somewhat surprisingly, there is no function for integer powers in the OCaml standard library (nor the course standard library.

Implement the function pow which, given an integer n and a nonnegative integer k, returns the integer n^k.

let _ =
  let _ = assert (pow 2 3 = 8) in
  let _ = assert (pow (-2) 3 = -8) in
  let _ = assert (pow 3 4 = 81) in
  ()

The behavior of the implementation is undefined if the given exponent is negative.

Put your solution in the file assign01/lib/assign01_01.ml. See the file assign01/test/test_suite/test_01.ml for example output behavior.