Module Stdlib320.Char

val lowercase_ascii : char -> char

lowercase_ascii c is the lowercase form of c if c is a letter, and is c otherwise.

let _ =
  let _ = assert (lowercase_ascii 'A' = 'a') in
  let _ = assert (lowercase_ascii 'a' = 'a') in
  let _ = assert (lowercase_ascii '2' = '2') in
  ()

val uppercase_ascii : char -> char

uppercase_ascii c is the uppercase form of c if c is a letter, and is c otherwise.

let _ =
  let _ = assert (lowercase_ascii 'a' = 'A') in
  let _ = assert (lowercase_ascii 'A' = 'A') in
  let _ = assert (lowercase_ascii '2' = '2') in
  ()