Module Specifications

module Assign0 : sig ... end

This assignment is due never is not graded. It's a practice run so you can verify that everything is set up correctly on your machine, and you can submit to Gradescope.