CAS CS 320: Concepts of Programming Languages

What is this course?

CAS CS 320 is a course about programming languages, particularly the design and implementation of programming languages. In this course, we take up the "programming language" as an object of formal study. This course is not about how to program, though the principles we cover are generally useful for designing and reasoning about programs.

The first part of the course is on functional programming in OCaml, based on CS 3110: Data Structures and Functional Programming at Cornell University and its associated textbook (referred to as OCP below). Its topics include inductive types, data abstraction, higher-order functions, combinators, and polymorphism. It is during this part that we learn to think functionally, to view programs not (necessarily) as sequences of commands manipulating some global state, but as compositions of (pure) functions which decompose and reorient data.

In the second part of the course we implement an interpreter for a simple (but ultimately quite powerful) programming language. The topics covered include parsing, syntax and semantics, variable scope and binding, storage management, and infinite data types. After this part, you will have your own implementation of a programming language for which you can write and run programs (pretty neat).

Prerequisites

CAS CS 111 & 112: Introduction to Computer Science

CAS CS 131: Combinatoric Structures

CAS CS 210: Computer Systems

Generally speaking, you will get the most out of this course (and will have the best time of it) if you complete the 200 level courses required for the degree of computer science beforehand. Above all, experience with a high-level programming language (like Python or Java) is essential.

Requisites

The workload for this course consists of assignments, a project, a midterm exam, and a final exam. Your final grade in the course will be determined by the following breakdown:

30% Assignments (6 of 8 assignments, 5% each)

20% Programming project (3 parts, ≈7% each)

20% Midterm exam

30% Final exam

Minutiae

Instructors: Assaf Kfoury and Nathan Mull

Teaching Fellows: Zachery Casey and Qiancheng (Robin) Fu

Course Assistants: Jason Wang and Sebastian Wu

Lecture Times: Tuesday and Thursdays, 3:30PM-4:45PM

Lecture Location: Stone Science Auditorium (STO) B-50

Lab Times/Locations: See the registrar page

See the Course Information section for more information.

Important Links

Piazza

Gradescope

Course Repository

Office hours calendar

Full List of Topics

OCaml

VS Code

Schedule

table {tr {th Date th Topic th Reading th Assignments

}

tr {td 1/18 td {{!page-topics.course_intro Course Introduction

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/intro/intro.html 1.(1-3)

}

}

td HW0 out

}

tr {td 1/23 td {{!page-topics.ocaml_intro_i Beginning Ocaml I: Expressions

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/basics/toplevel.html 2.1

}

, 2.2.(1-2), 2.3

}

td

}

tr {td 1/24 td Week 2 Lab: Set-Up Help td td

}

tr {td 1/25 td {{!page-topics.ocaml_intro_ii Beginning OCaml II: Functions

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/basics/functions.html 2.4

}

, 2.6

}

td HW1 out, HW0 due

}

tr {td 1/30 td {{!page-topics.ocaml_intro_iii Beginning OCaml III: Unions, Products

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/data/variants.html 3.2

}

, 3.4, 3.9.(1-3)

}

td

}

tr {td 1/31 td Week 3 Lab td td

}

tr {td 2/1 td {{!page-topics.ocaml_intro_iv Begnning Ocaml IV: Lists, IO

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/basics/printing.html 2.6

}

, 3.1

}

td HW2 out, HW1 due

}

tr {td 2/6 td {{!page-topics.ind_types_i Inductive Types I: Introduction

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/data/algebraic_data_types.html 3.9

}

}

td

}

tr {td 2/7 td Week 4 Lab td td

}

tr {td 2/8 td {{!page-topics.ind_types_ii Inductive Types II: Examples

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/data/algebraic_data_types.html 3.9

}

, 3.11

}

td HW3 out, HW2 due

}

tr {td 2/13 td {b Class canceled

}

td td

}

tr {td 2/14 td Week 5 Lab td td

}

tr {td 2/15 td {{!page-topics.ho_i Higher-Order Programming I: Maps, Filters

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/hop/intro.html 4.(1-3)

}

}

td HW4 out, HW3 due

}

tr {td 2/20 td {{!page-topics.ho_ii Higher-Order Programming II: Folds, Examples

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/hop/intro.html 4.(4-7)

}

}

td

}

tr {td 2/21 td {b Substitute Monday (No Lab)

}

td td

}

tr {td 2/22 td {{!page-topics.inter_i Intermediate OCaml I: Options, Monads, Modules

}

}

td OCP {{: https://cs3110.github.io/textbook/chapters/data/options.html 3.7

}

, 8.7, 5.1, 5.2

}

td HW4 due

}

tr {td 2/27 td {b Midterm Exam

}

td td

}

tr {td 2/28 td Week 7 Lab td td

}

tr {td 2/29 td {{!page-topics.grammar_i Formal Grammar I: Introduction

}

}

td td HW5 out

}

tr {td 3/5 td {{!page-topics.grammar_ii Formal Grammar II: Ambiguity

}

}

td td

}

tr {td 3/6 td Week 8 Lab td td

}

tr {td 3/7 td {{!page-topics.grammar_iii Formal Grammar III: Extended BNF, Regular Expressions

}

}

td td HW6 out, HW5 due

}

tr {td td {b Spring Recess

}

td td

}

tr {td 3/19 td {{!page-topics.parse_i Parsing I: Introduction

}

}

td td

}

tr {td 3/20 td Week 9 Lab td td

}

tr {td 3/21 td {{!page-topics.parse_ii Parsing II: General Parsing

}

}

td td HW7 out, HW6 due

}

tr {td 3/26 td {{!page-topics.parse_iii Parsing III: Combinators

}

}

td td

}

tr {td 3/27 td Week 10 Lab td td

}

tr {td 3/28 td {{!page-topics.sem_i Formal Semantics I: Operational Semantics

}

}

td td HW8 out, HW7 due

}

tr {td 4/2 td {{!page-topics.sem_ii Formal Semantics II: Applying Rules

}

}

td td

}

tr {td 4/3 td Week 11 Lab td td

}

tr {td 4/4 td {{!page-topics.sem_iii Formal Semantics III: Designing Rules

}

}

td td P1 out, HW8 due

}

tr {td 4/9 td {{!page-topics.sub_i Subprograms I: Variable Binding and Scoping

}

}

td td

}

tr {td 4/10 td Week 12 Lab td td

}

tr {td 4/11 td {{!page-topics.sub_ii Subprograms II: Parameter Passing

}

}

td td P2 out, P1 due

}

tr {td 4/16 td {{!page-topics.sub_iii Subprograms III: Activations Records and Closures

}

}

td td

}

tr {td 4/17 td Week 13 Lab td td

}

tr {td 4/18 td {{!page-topics.sub_iv Subprograms IV: The Substitution Model

}

}

td td P3 out, P2 due

}

tr {td 4/23 td {{!page-topics.adv_topic_i Advanced Topic: Compilation

}

}

td td

}

tr {td 4/24 td Week 14 Lab td td

}

tr {td 4/25 td {{!page-topics.adv_topic_ii Advanced Topic: The Lean Theorem Prover

}

}

td td P3 due

}

tr {td 4/30 td Final exam review session td td

}

}

Course Information

Lectures

Lectures are held twice a week, and the two instructors of the course will trade-off teaching one of the two lectures each week.

Labs

Labs will be held weekly by our TFs and TAs. More detail on the nature of these labs to come.

Communication

Announcements and discussions will happen on Piazza. If you are unfamiliar with Piazza, see their support page for information and tutorials. Some policies regarding the use of Piazza:

Don’t ask homework questions directly. Formulate a question which will aid your understanding, and could potentially help others as well.

Don’t give homework solutions directly. The answer-provider is just as culpable as the answer-receiver in cases of academic misconduct.

Piazza is as useful as it is active. We will try to answer questions on Piazza in a timely manner, but don’t hesitate to answer questions yourself.

You will also be expected to periodically check this webpage for updates.

Material

The first part of the course depends on the textbook OCaml Programming: Correct + Efficient + Beautiful.

All other course material will be made available on the courses GitHub repository (more information on this shortly). If you are unfamiliar with git and GitHub, see the GitHub documentation for information and tutorials. We will also cover some basics in labs.

Assignments

There will be 8 graded assignments throughout the semester. They are made available on Thursdays and are due the following Thursday by 11:59PM (see the Schedule above). Roughly half are written assignments and half are programming assignments.

Assignments will be submitted via Gradescope. If you are unfamiliar with Gradescope, see their Get Started page for more information. We will drop your lowest two assignment grades so only 6 assignments will count towards your final grade in the course. This means we will not accept late homework assignments, under any circumstances.

Final Project

Part of your final grade in the course will be determined by your completion of a final project. The project has three parts which makes up the final three-weeks-worth of assignments. More information about the nature of this project will be made available in the second-half of the course.

Policies and Guidelines

Please read carefully through my general BU Course Policies page. We take all matters of student mental health and safety, inclusion, and academic misconduct extremely seriously.

Log

1/25 Updated structure of site, topics now appear on their own page

12/31 Site is drafted