Module Specifications

module Assign0 : sig ... end

module Assign1 : sig ... end

module Assign2 : sig ... end

module Assign3 : sig ... end

module Assign4 : sig ... end

module Lab3 : sig ... end

module Lab4 : sig ... end

module Lab5 : sig ... end

Higher-order functions are particularly useful when we have to do a lot of list processing, e.g., for many applications in scheduling. In this lab, we'll be building a toy-version of when2meet, an online tool for determining a meeting time against a collection of potentially conflicting schedules.

module Stdlib320 : sig ... end