Module Stdlib320.Random

A module containing basic operations for randomness

val int : int -> int

int n is a random integer chosen uniformly from the range [0, n)

val float : float -> float

float n is a random float chosen from the range [0, n)

val bool : unit -> bool

bool () is a random Boolean value