Module Specifications

module Assign0 : sig ... end

module Assign1 : sig ... end

module Assign2 : sig ... end

module Assign3 : sig ... end

module Assign4 : sig ... end

module Assign5 : sig ... end

module Assign6 : sig ... end

This assignment is due on Thursday 10/23 by 8:00PM. You should put all of your programming solutions in a file called assign6/lib/assign6.ml. See the file test/test_assign6.ml for example behavior of each function. You should put all your written solutions in a single pdf file.