Module Specifications

module Assign0 : sig ... end

This assignment is due never is not graded. It's a practice run so you can verify that everything is set up correctly on your machine, and you can submit to Gradescope.

module Assign1 : sig ... end

This assignment is due on Thursday 9/11 by 8:00PM. You should put all of your solutions in a file called assign1/lib/assign1.ml. See the file test/test_assign1.ml for example behavior of each function.

module Assign2 : sig ... end