Module Specifications

module Assign0 : sig ... end

This assignment is due never is not graded. It's a practice run so you can verify that everything is set up correctly on your machine, and you can submit to Gradescope.

module Assign1 : sig ... end

This assignment is due on Thursday 9/11 by 8:00PM. You should put all of your solutions in a file called assign1/lib/assign1.ml. See the file test/test_assign1.ml for example behavior of each function.

module Assign2 : sig ... end

This assignment is due on Thursday 9/18 by 8:00PM. You should put all of your programming solutions in a file called assign2/lib/assign2.ml. See the file test/test_assign2.ml for example behavior of each function. You should put all your written solutions in a single pdf file.

module Assign3 : sig ... end

This assignment is due on Thursday 9/26 by 8:00PM. You should put all of your programming solutions in a file called assign3/lib/assign3.ml. See the file test/test_assign3.ml for example behavior of each function. You should put all your written solutions in a single pdf file.

module Assign4 : sig ... end

This assignment is due on Thursday 10/2 by 8:00PM. You should put all of your programming solutions in a file called assign4/lib/assign4.ml. See the file test/test_assign4.ml for example behavior of each function. You should put all your written solutions in a single pdf file.

module Assign5 : sig ... end

This assignment is due on Thursday 10/9 by 8:00PM. You should put all of your programming solutions in a file called assign5/lib/assign5.ml. See the file test/test_assign5.ml for example behavior of each function. You should put all your written solutions in a single pdf file.