CAS CS 320: Concepts of Programming Languages

What is this course?

CAS CS 320 is a course about programming languages, particularly the design and implementation of programming languages. In this course, we take up the "programming language" as an object of formal study. This course is not about how to program, though the principles we cover are generally useful for designing and reasoning about programs.

The first part of the course is on functional programming in OCaml, based on CS 3110: Data Structures and Functional Programming at Cornell University and its associated textbook (referred to as OCP below). Its topics include inductive types, data abstraction, higher-order functions, combinators, and polymorphism. It is during this part that we learn to think functionally, to view programs not (necessarily) as sequences of commands manipulating some global state, but as compositions of (pure) functions which decompose and reorient data.

In the second part of the course we implement an interpreter for a simple (but ultimately quite powerful) programming language. The topics covered include parsing, syntax and semantics, variable scope and binding, storage management, and infinite data types. After this part, you will have your own implementation of a programming language for which you can write and run programs (pretty neat).

Prerequisites

CAS CS 111 & 112: Introduction to Computer Science

CAS CS 131: Combinatoric Structures

CAS CS 210: Computer Systems

Generally speaking, you will get the most out of this course (and will have the best time of it) if you complete the 200 level courses required for the degree of computer science beforehand. Above all, experience with a high-level programming language (like Python or Java) is essential.

Requisites

The workload for this course consists of assignments, a project, a midterm exam, and a final exam. Your final grade in the course will be determined by the following breakdown:

30% Assignments (6 of 8 assignments, 5% each)

20% Programming project (3 parts, ≈7% each)

20% Midterm exam

30% Final exam

Minutiae

Instructors: Assaf Kfoury and Nathan Mull

Teaching Fellows: Zachery Casey and Qiancheng (Robin) Fu

Course Assistants: Jason Wang and Sebastian Wu

Lecture Times: Tuesday and Thursdays, 3:30PM-4:45PM

Lecture Location: Stone Science Auditorium (STO) B-50

Lab Times/Locations: See the registrar page

See the syllabus for this course for more logistical details regarding the course.

Important Links

Piazza

Gradescope

Office hours calendar

Course Information

OCaml

VS Code

Schedule

Date

Topic

Reading

Assignments

1/18

What is this course?

OCP 1.(1-3)

HW0 out

1/23

Beginning OCaml I: Expressions

OCP 2.1, 2.2.(1-2), 2.3

1/24

Week 2 Lab: Set-Up Help

1/25

Beginning OCaml II: Functions

OCP 2.4, 2.6

HW1 out, HW0 due

1/30

Beginning OCaml III: Unions, products

OCP 3.2, 3.4

1/31

Week 3 Lab

2/1

Beginning OCaml IV: IO, Lists

OCP 3.1

HW2 out, HW1 due

2/6

Inductive types I: Introduction

OCP 3.9.(1-4)

2/7

Week 4 Lab

2/8

Inductive types II: Examples

OCP 3.9.(5-7), 3.11

HW3 out, HW2 due

2/13

Higher-Order Programming I: Map, filter

OCP 4.(1-3)

2/14

Week 5 Lab

2/15

Higher-Order Programming II: Fold, examples

OCP 4.(4-7)

HW4 out, HW3 due

2/20

Intermediate OCaml I: Options, monads

OCP 3.7, 8.7

2/21

Substitute Monday

2/22

Intermediate OCaml II: Modules

OCP 5.1, 5.2

HW4 due

2/27

Midterm Exam

2/28

Week 7 Lab

2/29

Formal Grammar I: Introduction

HW5 out

3/5

Formal Grammar II: Precedence

3/6

Week 8 Lab

3/7

Formal Grammar III: Extended BNFs, regular expressions

HW6 out, HW5 due

Spring Recess

3/19

Parsing I: Introduction

3/20

Week 9 Lab

3/21

Parsing II: Parsing in OCaml

HW7 out, HW6 due

3/26

Parsing III: Combinators

3/27

Week 10 Lab

3/28

Formal Semantics I: Operational Semantics

HW8 out, HW7 due

4/2

Formal Semantics II: Applying Rules

4/3

Week 11 Lab

4/4

Formal Semantics III: Designing Rules

P1 out, HW8 due

4/9

Variables I: Scoping

4/10

Week 12 Lab

4/11

Variables II: Binding

P2 out, P1 due

4/16

Subprograms I: Introduction

4/17

Week 13 Lab

4/18

Subprograms II: Activation records

P3 out, P2 due

4/23

Subprograms III: Activation stack

4/24

Week 14 Lab

4/25

Advanced Topics I: Infinite Data Types

P3 due

4/30

Final exam review session

Topics

What is this course?

Overview: We discuss the expectations of the course, and give a rough outline of the topics that will be covered. We will also touch on the benefits of the functional paradigm.

Keywords: course structure, course policy, OCaml, functional programming

Required reading: OCaml Programming (OCP) 1.1-1.3: Better programming through OCaml

Additional information: Homework 0 is released. It's not graded but it's important that you complete it so that you're set up to submit the later assignments.

Beginning OCaml

I: Expressions

Overview: We look at a couple ways of interacting with OCaml via UTop. We discuss expressions in OCaml, particularly if-expressions and let-expressions, and begin looking more careful at evaluation.

Keywords: UTop, values, expressions, evaluation, types, syntax, dynamic semantics, static semantics, if-expressions, let-expressions, let-defintions

Required reading:

OCP 2.1: The OCaml Toplevel

OCP 2.2.(1-2): Compiling OCaml Programs

OCP 2.3: Expressions

Additional information:

Homework 0 is due on Thursday by 11:59PM

The office hours calendar has been updated with locations

Attend lab this week if you need help setting up your programming environment

II: Functions

III: Unions, products

IV: IO, Lists

Inductive types

I: Introduction

II: Examples

Higher-Order Programming

I: Map, filter

II: Fold, examples

Intermediate OCaml

I: Options, monads

II: Modules

Formal Grammar

I: Introduction

II: Precedence

III: Extended BNFs, regular expressions

Parsing

I: Introduction

II: Parsing in OCaml

III: Combinators

Formal Semantics

I: Operational Semantics

II: Applying Rules

III: Designing Rules

Variables

I: Scoping

II: Binding

Subprograms

I: Introduction

II: Activation records

III: Activation stack

Advanced Topics

I: Infinite Data Types

Final Exam Review

Course Information

Lectures

Lectures are held twice a week, and the two instructors of the course will trade-off teaching one of the two lectures each week.

Labs

Labs will be held weekly by our TFs and TAs. More detail on the nature of these labs to come.

Communication

Announcements and discussions will happen on Piazza. If you are unfamiliar with Piazza, see their support page for information and tutorials. Some policies regarding the use of Piazza:

Don’t ask homework questions directly. Formulate a question which will aid your understanding, and could potentially help others as well.

Don’t give homework solutions directly. The answer-provider is just as culpable as the answer-receiver in cases of academic misconduct.

Piazza is as useful as it is active. We will try to answer questions on Piazza in a timely manner, but don’t hesitate to answer questions yourself.

You will also be expected to periodically check this webpage for updates.

Material

The first part of the course depends on the textbook OCaml Programming: Correct + Efficient + Beautiful.

All other course material will be made available on the courses GitHub repository (more information on this shortly). If you are unfamiliar with git and GitHub, see the GitHub documentation for information and tutorials. We will also cover some basics in labs.

Assignments

There will be 8 graded assignments throughout the semester. They are made available on Thursdays and are due the following Thursday by 11:59PM (see the Schedule above). Roughly half are written assignments and half are programming assignments.

Assignments will be submitted via Gradescope. If you are unfamiliar with Gradescope, see their Get Started page for more information. We will drop your lowest two assignment grades so only 6 assignments will count towards your final grade in the course. This means we will not accept late homework assignments, under any circumstances.

Final Project

Part of your final grade in the course will be determined by your completion of a final project. The project has three parts which makes up the final three-weeks-worth of assignments. More information about the nature of this project will be made available in the second-half of the course.

Policies and Guidelines

Please read carefully through my general BU Course Policies page. We take all matters of student mental health and safety, inclusion, and academic misconduct extremely seriously.

Log

1/31 Site is drafted