CAS CS 320: Concepts of Programming Languages

What is this course?

CAS CS 320 is a course about programming languages, particularly the design and implementation of programming languages. In this course, we take up the "programming language" as an object of formal study. This course is not about how to program, though the principles we cover are generally useful for designing and reasoning about programs.

The first part of this course is on functional programming in OCaml, based on CS 3110: Data Structures and Functional Programming at Cornell University and its associated textbook (referred to as OCP below). Its topics include strong typing, data abstraction, higher-order functions, combinators, and polymorphism. It is during this part that we learn to think functionally, to view programs not (necessarily) as sequences of commands manipulating some global state, but as compositions of (pure) functions which decompose and reorient data.

In the second part of the course we implement an interpreter for a simple (but ultimately quite powerful) programming language. The part includes topics like include parsing, syntax and semantics, variable scope and binding, storage management, and infinite data types. After this part, you will have your own implementation of a programming language for which you can write and run programs (pretty neat).

Prerequisites

CAS CS 111 & 112: Introduction to Computer Science

CAS CS 131: Combinatoric Structures

CAS CS 210: Computer Systems

Generally speaking, you will get the most out of this course (and will have the best time of it) if you complete the 200 level courses required for the degree of computer science beforehand. Above all, experience with a high-level programming language (like Python or Java) is essential.

Requisites

The workload for this course consists of assignments, a project, and a final exam. Your final grade in this course will be determined by the following breakdown:

20% Written assignments (4 assignments, 5% each)

20% Programming assignments (4 assignments, 5% each)

30% Programming project (3 parts, 10% each)

30% Final exam

Course Staff

Instructors: Assaf Kfoury and Nathan Mull

Teaching Fellows/Assistants: TBD

Course Assistants: TBD

See the syllabus for this course for more logistical details regarding the course.

Important Links

Piazza

Gradescope

Office hours calendar

OCaml

Syllabus

Schedule

Week

Date

Topic

Reading

Notes

1

1/18

What is this course?

OCP 1.1-1.3

2

1/23

Beginning OCaml I: Expressions

OCP 2.1, 2.2.(1-2), 2.3

1/24

Lab 1: TBD

1/25

Beginning OCaml II: Functions

OCP 2.4, 2.6

3

1/30

Beginning OCaml III: Unions, products

OCP 3.2, 3.4

1/31

Lab 2: TBD

2/1

Beginning OCaml IV: IO, Lists

OCP 3.1

4

2/6

Inductive types I: Introduction

OCP 3.9.(1-4)

2/7

Lab 3: TBD

2/8

Inductive types II: Examples

OCP 3.9.(5-7), 3.11

5

2/13

Higher-Order Programming I: Map, filter

OCP

2/14

Lab 4: TBD

2/15

Higher-Order Programming II: Fold, examples

6

2/20

Intermediate OCaml I: Options, monads

OCP 3.7, OCP 8.7

2/21

Substitute Monday

2/22

Intermediate OCaml II: Modules

7

2/27

Intermediate OCaml III: Mutability

2/28

Lab 5: TBD

2/29

Formal Grammar I: Introduction

8

3/5

Formal Grammar II: Ambiguity, precedence

3/6

Lab 6: TBD

3/7

Formal Grammar III: Extended BNFs, regular expressions

Spring Recess

9

3/19

Parsing I: Introduction

3/20

Lab 7: TBD

3/21

Parsing II: Combinators

10

3/26

Formal Semantics I: Operational Semantics

3/27

Lab 8: TBD

3/28

Formal Semantics II: Rules and Judgments

11

4/2

Variables I: Scoping

4/3

Lab 9: TBD

4/4

Variables II: Binding

12

4/9

Subprograms I: Introduction

4/10

Lab 10: TBD

4/11

Subprograms II: Activation records

13

4/16

Subprograms III: Activation stack

4/17

Lab 11: TBD

4/18

Infinite data types I: Introduction

14

4/23

Infinite data types II: Streams

4/24

Lab 12: TBD

4/25

Infinite data types III: Examples of streams

15

4/30

Final exam review session

Material

Slides

Labs

Topics

What is this course?

This is a very simple summary of what is covered in this lecture. It's not meant to be exhaustive, just make more clear what it is that we are supposed to cover. What are the learning objectives.

Required reading: OCaml Programming (OCP) 1.1-1.3: Better programming through OCaml

Learning objectives

Beginning OCaml

I: Expressions

II: Functions

III: Unions, products

IV: IO, Lists

Inductive types

I: Introduction

II: Examples

Higher-Order Programming

I: Map, filter

II: Fold, examples

Intermediate OCaml

I: Options, monads

II: Modules

III: Mutability

Formal Grammar

I: Introduction

II: Ambiguity, precedence

III: Extended BNFs, regular expressions

Parsing

I: Introduction

II: Combinators

Formal Semantics

I: Operational Semantics

II: Rules and Judgments

Variables

I: Scoping

II: Binding

Subprograms

I: Introduction

II: Activation records

III: Activation stack

Infinite data types

I: Introduction

II: Streams

III: Examples of streams

Final Exam Review

Log