Module Specifications

module Assign0 : sig ... end

This assignment is due on Thursday 1/23 by 8:00PM and is not graded. It's a practice run so you can verify that everything is set up correctly on your machine, and you can submit to Gradescope.

module Stdlib320 : sig ... end

The following is the standard library we'll use for the first half of CAS CS 320: Concepts of Programming Languages at Boston University (Spring 2025). It's a very small subset of the OCaml Standard Library with a bit more documentation (it also gives us more control over what functions are "allowed" for assignments). Nearly everything here is also available in the OCaml Standard Library, so check the Stdlib documentation for more information on what you see below.